+++ src/main/scala/com/dataintuitive/viash/platforms/DockerPlatform.scala
@@ -27,6 +27,9 @@ import com.dataintuitive.viash.config.Version
 import com.dataintuitive.viash.wrapper.{BashWrapper, BashWrapperMods}
 import com.dataintuitive.viash.platforms.docker._
 import com.dataintuitive.viash.helpers.Circe._
+import com.dataintuitive.viash.config.Info
+import java.util.Date
+import java.text.SimpleDateFormat
 
 case class DockerPlatform(
   id: String = "docker",
@@ -90,7 +93,7 @@ case class DockerPlatform(
       { if (privileged) " --privileged" else "" }
 
     // create setup
-    val (effectiveID, setupMods) = processDockerSetup(functionality, testing)
+    val (effectiveID, setupMods) = processDockerSetup(functionality, config.info, testing)
 
     // generate automount code
     val dmVol = processDockerVolumes(functionality)
@@ -136,21 +139,32 @@ case class DockerPlatform(
     )
   }
 
-  private def processDockerSetup(functionality: Functionality, testing: Boolean) = {
+  private def processDockerSetup(functionality: Functionality, info: Option[Info], testing: Boolean) = {
     // construct labels from metadata
-    val auth = functionality.authors match {
-      case Nil => Nil
-      case aut: List[Author] => List(s"""authors="${aut.mkString(", ")}"""")
+    val opencontainers_image_authors = functionality.authors match {
+      case Nil => None
+      case aut: List[Author] => Some(aut.mkString(", "))
     }
-    /*val descr = functionality.description.map{ des => 
-      s"""org.opencontainers.image.description="${Bash.escape(des)}""""
-    }.toList*/
-    
-    val descr = List(
-      s"""org.opencontainers.image.description="Companion container for running component ${functionality.namespace.map(_ + " ").getOrElse("")}${functionality.name}""""
-    )
-    val imageSource = target_image_source.map(des => s"""org.opencontainers.image.source="${Bash.escape(des)}"""").toList
-    val labelReq = DockerRequirements(label = auth ::: descr ::: imageSource)
+    // if no target_image_source is defined,
+    // translate git@github.com:viash-io/viash.git -> https://github.com/viash-io/viash.git
+    val opencontainers_image_source = (target_image_source, info) match {
+      case (Some(tis), _) => Some(tis)
+      case (None, Some(i)) =>
+        i.git_remote.map(url => url.replaceAll(":([^/])", "/$1").replaceAllLiterally("ssh//", "").replaceAllLiterally("git@", "https://"))
+      case _ => None
+    }
+    val opencontainers_image_revision = info.flatMap(_.git_commit)
+    val opencontainers_image_version = info.flatMap(_.git_tag)
+    val opencontainers_image_description = s""""Companion container for running component ${functionality.namespace.map(_ + " ").getOrElse("")}${functionality.name}""""
+    val opencontainers_image_created = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssXXX").format(new Date())
+
+    val authors = opencontainers_image_authors.map(aut => s"""org.opencontainers.image.authors="$aut"""").toList
+    val descr = List(s"org.opencontainers.image.description=$opencontainers_image_description")
+    val imageSource = opencontainers_image_source.map(des => s"""org.opencontainers.image.source="${Bash.escape(des)}"""").toList
+    val created = List(s"""org.opencontainers.image.created="$opencontainers_image_created"""")
+    val revision = opencontainers_image_revision.map(rev => s"""org.opencontainers.image.revision="$rev"""").toList
+    val version = opencontainers_image_version.map(v => s"""org.opencontainers.image.version="$v"""").toList
+    val labelReq = DockerRequirements(label = authors ::: descr ::: created ::: imageSource ::: revision ::: version)
 
     val setupRequirements = testing match {
       case true => test_setup